singer-connectors/tap-kafka/requirements.txt
python-snappy
kafka-python
singer-python==5.2.0
requests==2.12.4
psycopg2==2.7.4
strict-rfc3339==0.7
nose==1.3.7
jsonschema==2.6.0